課題２の修正
