+++ src/types/Post.ts
@@ -1,0 +1,10 @@
+import { Category } from "../post/Categories";
+
+export type Post = {
+  id: number;
+  title: string;
+  thumbnailUrl: string;
+  createdAt: string;
+  categories: string[];
+  content: string;
+};
